# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: StayHub v2 API
  version: 1.0.0
  description: |
    API contract for StayHub.

    This `openapi.yaml` is intended as the implementation target for the checklist in `task.md`.
servers:
  - url: http://localhost:8080
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/bookings/book:
    post:
      tags: [Bookings]
      summary: Create a new accommodation reservation
      operationId: bookAccommodation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created (includes deposit payment summary)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/bookings/{reservationId}:
    get:
      tags: [Bookings]
      summary: Get reservation detail by ID
      description: Only the guest or the host of the accommodation can access it.
      operationId: getReservationById
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/bookings/my-reservations:
    get:
      tags: [Bookings]
      summary: Get paginated list of my reservations
      description: If HOST returns reservations of owned accommodations, else guest reservations.
      operationId: getMyReservations
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Reservations page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageReservationSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v2/bookings/{reservationId}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel a reservation
      description: Cancels an ACTIVE reservation. Allowed for the guest or the host.
      operationId: cancelReservation
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelReservationRequest'
      responses:
        '200':
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/bookings/{reservationId}/reschedule:
    patch:
      tags: [Bookings]
      summary: Reschedule a reservation
      description: Updates the start and end dates of an ACTIVE reservation. Only the booking guest can reschedule.
      operationId: rescheduleReservation
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleReservationRequest'
      responses:
        '200':
          description: Reservation rescheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dates not available
  /api/v2/bookings/{reservationId}/payment-summary:
    get:
      tags: [Payments]
      summary: Get reservation payment summary
      operationId: getReservationPaymentSummary
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationPaymentSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/bookings/{reservationId}/payments/deposit:
    post:
      tags: [Payments]
      summary: Report deposit payment
      description: |
        Marks the reservation deposit as paid.
        This can be implemented as either an admin/host confirmation flow or an automated reconciliation.
      operationId: reportDepositPayment
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositPaymentReportRequest'
      responses:
        '200':
          description: Deposit payment registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationPaymentSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/bookings/{reservationId}/confirmation/resend:
    post:
      tags: [Bookings]
      summary: Resend reservation confirmation
      description: Triggers a new confirmation email for the reservation.
      operationId: resendReservationConfirmation
      parameters:
        - in: path
          name: reservationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Confirmation email triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Invalid login attempt detected
        code:
          type: integer
          example: 401
      required: [message, code]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operacion realizada exitosamente
      required: [message]
...
    ReservationStatus:
      type: string
      enum: [ACTIVE, COMPLETED, CANCELLED, NO_SHOW]
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      example: COP
...
    CreateReservationRequest:
      type: object
      description: Request body for creating a new reservation.
      properties:
        accommodationId:
          type: integer
          format: int64
          example: 1
        startDate:
          type: string
          format: date-time
          example: '2026-05-10T15:00:00'
        endDate:
          type: string
          format: date-time
          example: '2026-05-12T11:00:00'
        roomCode:
          type: string
          nullable: true
          description: Optional room code if booking is room-specific.
      required: [accommodationId, startDate, endDate]
    CreateReservationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        bookingNumber:
          type: string
          example: SH-2026-000001
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalPrice:
          type: number
          format: double
          example: 1000.0
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        status:
          $ref: '#/components/schemas/ReservationStatus'
        accommodationId:
          type: integer
          format: int64
        accommodationTitle:
          type: string
        userId:
          type: integer
          format: int64
        depositAmount:
          type: number
          format: double
          example: 200.0
        depositPaid:
          type: boolean
          example: false
        bankAccountNumber:
          type: string
          example: '3001234567890'
        paymentDeadline:
          type: string
          format: date-time
      required:
        - id
        - bookingNumber
        - startDate
        - endDate
        - totalPrice
        - currency
        - status
        - accommodationId
        - accommodationTitle
        - userId
        - depositAmount
        - depositPaid
        - bankAccountNumber
        - paymentDeadline
    RetrieveReservationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingNumber:
          type: string
          example: SH-2026-000001
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        accommodationId:
          type: integer
          format: int64
        accommodationTitle:
          type: string
        accommodationCity:
          type: string
        guestId:
          type: integer
          format: int64
        guestEmail:
          type: string
          format: email
        totalPrice:
          type: number
          format: double
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        depositAmount:
          type: number
          format: double
        depositPaid:
          type: boolean
        paymentDeadline:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ReservationStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - bookingNumber
        - startDate
        - endDate
        - accommodationId
        - accommodationTitle
        - accommodationCity
        - guestId
        - guestEmail
        - totalPrice
        - currency
        - depositAmount
        - depositPaid
        - paymentDeadline
        - status
        - createdAt
        - updatedAt
    RetrieveReservationSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accommodationId:
          type: integer
          format: int64
        accommodationTitle:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        totalPrice:
          type: number
          format: double
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        status:
          $ref: '#/components/schemas/ReservationStatus'
      required: [id, accommodationId, accommodationTitle, startDate, endDate, totalPrice, currency, status]
    PageReservationSummary:
      type: object
      description: Spring Data Page of reservation summaries.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RetrieveReservationSummary'
        number:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 25
        totalPages:
          type: integer
          example: 3
        first:
          type: boolean
        last:
          type: boolean
      required: [content, number, size, totalElements, totalPages, first, last]
    CancelReservationRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 200
          nullable: true
    RescheduleReservationRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [startDate, endDate]
    ReservationPaymentSummary:
      type: object
      properties:
        reservationId:
          type: integer
          format: int64
        bookingNumber:
          type: string
        totalPrice:
          type: number
          format: double
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        depositAmount:
          type: number
          format: double
        depositPaid:
          type: boolean
        paymentDeadline:
          type: string
          format: date-time
        bankAccountNumber:
          type: string
        overdue:
          type: boolean
      required: [reservationId, totalPrice, currency, depositAmount, depositPaid, paymentDeadline, bankAccountNumber, overdue]
    DepositPaymentReportRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
          example: 200.0
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        reference:
          type: string
          nullable: true
          example: TRX-009182
        paidAt:
          type: string
          format: date-time
          nullable: true
        proofUrl:
          type: string
          format: uri
          nullable: true
      required: [amount, currency]
...